@@ -2,17 +2,15 @@
 
 namespace App\Actions\Jetstream;
 
+use App\Models\User;
 use Laravel\Jetstream\Contracts\DeletesUsers;
 
 class DeleteUser implements DeletesUsers
 {
     /**
      * Delete the given user.
-     *
-     * @param  mixed  $user
-     * @return void
      */
-    public function delete($user)
+    public function delete(User $user): void
     {
         $user->deleteProfilePhoto();
         $user->tokens->each->delete();
