@@ -19,15 +19,15 @@ return [
     'stack' => 'livewire',
 
     /*
-     |--------------------------------------------------------------------------
-     | Jetstream Route Middleware
-     |--------------------------------------------------------------------------
-     |
-     | Here you may specify which middleware Jetstream will assign to the routes
-     | that it registers with the application. When necessary, you may modify
-     | these middleware; however, this default value is usually sufficient.
-     |
-     */
+    |--------------------------------------------------------------------------
+    | Jetstream Route Middleware
+    |--------------------------------------------------------------------------
+    |
+    | Here you may specify which middleware Jetstream will assign to the routes
+    | that it registers with the application. When necessary, you may modify
+    | these middleware; however, this default value is usually sufficient.
+    |
+    */
 
     'middleware' => ['web'],
 
