@@ -1,4 +1,4 @@
-<x-jet-action-section>
+<x-action-section>
     <x-slot name="title">
         {{ __('Two Factor Authentication') }}
     </x-slot>
@@ -38,7 +38,7 @@
                     </p>
                 </div>
 
-                <div class="mt-4">
+                <div class="mt-4 p-2 inline-block bg-white">
                     {!! $this->user->twoFactorQrCodeSvg() !!}
                 </div>
 
@@ -50,13 +50,13 @@
 
                 @if ($showingConfirmation)
                     <div class="mt-4">
-                        <x-jet-label for="code" value="{{ __('Code') }}" />
+                        <x-label for="code" value="{{ __('Code') }}" />
 
-                        <x-jet-input id="code" type="text" name="code" class="block mt-1 w-1/2" inputmode="numeric" autofocus autocomplete="one-time-code"
-                            wire:model.defer="code"
+                        <x-input id="code" type="text" name="code" class="block mt-1 w-1/2" inputmode="numeric" autofocus autocomplete="one-time-code"
+                            wire:model.live="code"
                             wire:keydown.enter="confirmTwoFactorAuthentication" />
 
-                        <x-jet-input-error for="code" class="mt-2" />
+                        <x-input-error for="code" class="mt-2" />
                     </div>
                 @endif
             @endif
@@ -78,47 +78,47 @@
 
         <div class="mt-5">
             @if (! $this->enabled)
-                <x-jet-confirms-password wire:then="enableTwoFactorAuthentication">
-                    <x-jet-button type="button" wire:loading.attr="disabled">
+                <x-confirms-password wire:then="enableTwoFactorAuthentication">
+                    <x-button type="button" wire:loading.attr="disabled">
                         {{ __('Enable') }}
-                    </x-jet-button>
-                </x-jet-confirms-password>
+                    </x-button>
+                </x-confirms-password>
             @else
                 @if ($showingRecoveryCodes)
-                    <x-jet-confirms-password wire:then="regenerateRecoveryCodes">
-                        <x-jet-secondary-button class="mr-3">
+                    <x-confirms-password wire:then="regenerateRecoveryCodes">
+                        <x-secondary-button class="me-3">
                             {{ __('Regenerate Recovery Codes') }}
-                        </x-jet-secondary-button>
-                    </x-jet-confirms-password>
+                        </x-secondary-button>
+                    </x-confirms-password>
                 @elseif ($showingConfirmation)
-                    <x-jet-confirms-password wire:then="confirmTwoFactorAuthentication">
-                        <x-jet-button type="button" class="mr-3" wire:loading.attr="disabled">
+                    <x-confirms-password wire:then="confirmTwoFactorAuthentication">
+                        <x-button type="button" class="me-3" wire:loading.attr="disabled">
                             {{ __('Confirm') }}
-                        </x-jet-button>
-                    </x-jet-confirms-password>
+                        </x-button>
+                    </x-confirms-password>
                 @else
-                    <x-jet-confirms-password wire:then="showRecoveryCodes">
-                        <x-jet-secondary-button class="mr-3">
+                    <x-confirms-password wire:then="showRecoveryCodes">
+                        <x-secondary-button class="me-3">
                             {{ __('Show Recovery Codes') }}
-                        </x-jet-secondary-button>
-                    </x-jet-confirms-password>
+                        </x-secondary-button>
+                    </x-confirms-password>
                 @endif
 
                 @if ($showingConfirmation)
-                    <x-jet-confirms-password wire:then="disableTwoFactorAuthentication">
-                        <x-jet-secondary-button wire:loading.attr="disabled">
+                    <x-confirms-password wire:then="disableTwoFactorAuthentication">
+                        <x-secondary-button wire:loading.attr="disabled">
                             {{ __('Cancel') }}
-                        </x-jet-secondary-button>
-                    </x-jet-confirms-password>
+                        </x-secondary-button>
+                    </x-confirms-password>
                 @else
-                    <x-jet-confirms-password wire:then="disableTwoFactorAuthentication">
-                        <x-jet-danger-button wire:loading.attr="disabled">
+                    <x-confirms-password wire:then="disableTwoFactorAuthentication">
+                        <x-danger-button wire:loading.attr="disabled">
                             {{ __('Disable') }}
-                        </x-jet-danger-button>
-                    </x-jet-confirms-password>
+                        </x-danger-button>
+                    </x-confirms-password>
                 @endif
 
             @endif
         </div>
     </x-slot>
-</x-jet-action-section>
+</x-action-section>
