@@ -13,49 +13,49 @@ class UpdatePasswordTest extends TestCase
 {
     use RefreshDatabase;
 
-    public function test_password_can_be_updated()
+    public function test_password_can_be_updated(): void
     {
         $this->actingAs($user = User::factory()->create());
 
         Livewire::test(UpdatePasswordForm::class)
-                ->set('state', [
-                    'current_password' => 'password',
-                    'password' => 'new-password',
-                    'password_confirmation' => 'new-password',
-                ])
-                ->call('updatePassword');
+            ->set('state', [
+                'current_password' => 'password',
+                'password' => 'new-password',
+                'password_confirmation' => 'new-password',
+            ])
+            ->call('updatePassword');
 
         $this->assertTrue(Hash::check('new-password', $user->fresh()->password));
     }
 
-    public function test_current_password_must_be_correct()
+    public function test_current_password_must_be_correct(): void
     {
         $this->actingAs($user = User::factory()->create());
 
         Livewire::test(UpdatePasswordForm::class)
-                ->set('state', [
-                    'current_password' => 'wrong-password',
-                    'password' => 'new-password',
-                    'password_confirmation' => 'new-password',
-                ])
-                ->call('updatePassword')
-                ->assertHasErrors(['current_password']);
+            ->set('state', [
+                'current_password' => 'wrong-password',
+                'password' => 'new-password',
+                'password_confirmation' => 'new-password',
+            ])
+            ->call('updatePassword')
+            ->assertHasErrors(['current_password']);
 
         $this->assertTrue(Hash::check('password', $user->fresh()->password));
     }
 
-    public function test_new_passwords_must_match()
+    public function test_new_passwords_must_match(): void
     {
         $this->actingAs($user = User::factory()->create());
 
         Livewire::test(UpdatePasswordForm::class)
-                ->set('state', [
-                    'current_password' => 'password',
-                    'password' => 'new-password',
-                    'password_confirmation' => 'wrong-password',
-                ])
-                ->call('updatePassword')
-                ->assertHasErrors(['password']);
+            ->set('state', [
+                'current_password' => 'password',
+                'password' => 'new-password',
+                'password_confirmation' => 'wrong-password',
+            ])
+            ->call('updatePassword')
+            ->assertHasErrors(['password']);
 
         $this->assertTrue(Hash::check('password', $user->fresh()->password));
     }
